requirements.txt
astroid==3.0.3
click==8.1.7
dill==0.3.8
install==1.3.5
isort==5.13.2
mccabe==0.7.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg==3.1.17
pylint==3.0.3
python-dotenv==1.0.1
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
wcwidth==0.2.13
